insert CSS resource with viewlet instead of using portal_css.
This is important so that we can distinguish between rendering the CSS on
the Plone site and rendering it on other navigation roots (e.g. subsites)
which may have a different CSS (changed colors etc).

This distinction is not possible when using portal_css since the bundles
are heavily cached.