change caching to include full URL instead of path.
This separates caches for different virtual host configurations
or other sources.
This is important because the $portal-url variable may be used in SCSS,
resulting in full URLs to be included in the cached CSS.

Since users should always download the resources from the same URL base as
they visit the website we should never return a CSS cache generated with
another URL base.