Implement client-side caching.
The theming.css is now cached in the browser cached because we set cache
control headers.

In order to be able to invalidate the cache we add a cachekey param to
the css url, so that we just can change the param and thus the URL in
order to invalidate the cache.

The cachekey is generated per navigation root and contains the URL to
the navigation root, a portal_css bundle compilation state hash and the
modified date of the navigation root.
The cachekey changes when the portal_css bundles are recooked or when
the navigation root is changed.

When the portal_css debug mode is enabled, caching is disabled
altogether by not appending a cachekey param.

Removed cache-invalidation on shift-reload because any user can
invalidate the server-side cache and force a sass compilation, which is
in fact a DDoS vector.