add theming.css redirect on non-navroots
The theming.css may be requested on any context.
In order to make the client-side cache work efficiently we only want to
have theming.css-requests on navigation roots.

Therefore all requests on non-navigation-roots are redirected to the
next parent navigation root, adding the cachekey param.

This fixes a TinyMCE problem:
When "theming.css" is configured as TinyMCE content CSS resource, it is
requested relative to the current context, which breaks caching because
the URL always changes.