Fix an error when gzipping the CSS.
Depending on the generated CSS, Zope's GZIP compression will fail and
result in an invalid response.
In order to mitigate this issue we no longer use
``enableHTTPCompression``.

By configuring a plone.app.caching ruleset we can let plone.app.caching
handle the compression, when it is installed and configured.
It will also override the Cache-Control header.
In order to have a Cache-Control header when p.a.caching is not
installed or not configured, we still set it.