always include dynamic resources cache keys.
Respect dynamic resources cache keys in production too, because the SCSS
of those resources might change any time.

In development we include cache keys of all resources in order to track
changes in scss files, but this is not necessary in production because
we include the portal_css bundle hashes and the portal_css bundles
should be recooked on production updates.