Font sizes are not explicitely defined.
Because the headings are defined in `em`s the size changes when an parent element does define
a diffrent font size than the base. So to be sure we always have the same fontsize we need to
define the font-sizes in `rem`s which base on the body's font size.