Extend floatgrid mixin with `by-index` directive.
When using the floatgrid on a set of elements with mixed types or classes
the legacy `by-index` mode is recommended becuase the floatgrid will
apply using the `nth-child` selector.

When using the floatgrid on a set of elements with mixed types or classes
but you just want to make an effect on a subset with similar types or classes
the new not `by-index` mode is recommended.