Adjust overlay size for large and small overlay.
According to https://github.com/4teamwork/plonetheme.blueberry/pull/52#issuecomment-214168577
the large overlay is too small. So we decided to make the large overlay
almost fullscreen so that the user is still able to see a bit of the
page in the transparent background.

The large overlay does not scale the size until the mobile breakpoint has been
reached.

The small overlay is too small on tiny screen becuase the height is relative to the
screensize. So we dediced to set a fixed height so that the user is still able
to see the content of the overlay on tiny screens.
We height is based on the plone login form and the simplelayout delete confirmation.