Fix gradient mixin.
Current gradient mixin is not able to handle complex gradients such as radial
or repeating gradients.
So the mixin is going to be limited to linear gradients.
For other or more complex gradient other mixins should do the job.
Fix safari bug: transparent is not working in safari use rgba(255, 255,
255, 0) instead.