Fix caching for unpublished navigation roots.
plone.app.caching does not ever cache resources which are not public (View
for Anonymous).
This means that when using plone.app.caching we cannot cache the
generated CSS within unpublished navigation roots at all.
Since the CSS does not really contain any information about the
navigation root it is not a problem to cache it, and we should cache it
whenever possible since building the CSS is considered slow.

Therefore we no longer use plone.app.caching in order to set the
Cache-Control response headers. We now always do this manually.