Improve mimetype icons support for better Plone 5 support.
Up to now, ftw.theming used the "getIcon" catalog column (containing an
image path) for generating a HTML structure which can be styled with
CSS / font awesome icons.

This strategy does no longer work well in Plone 5 since the "getIcon"
metadata column is no longer used.

This change adds an additional strategy for when there is no "getIcon"
value on the brain.
The new strategy uses the "mime_type" column for generating a css
class. This strategy may not work well with Plone 4 / depending on the
configuration, that's why we no have two strategies.

Since the strategies generated varying css classes, the icons need to
be re-defined based on the "mime_type" value in addition to the
definition based on the "getIcon" filename.

The change includes icon definitons for the already styled mimetypes
and an update of the icons control panel so that it is easier to
control and manage the missing mimetype definitions.